refactor: move accounting::server_cost types from lib to wire crate
