+++ wire/src/accounting/mod.rs
@@ -1,5 +1,7 @@
 mod server_consumption;
+mod server_cost;
 mod server_state;
 
 pub use server_consumption::*;
+pub use server_cost::*;
 pub use server_state::*;
